#55 artice agg 추가하기 - jdbc config commons로 이동
